+++ ax25_pad.h
@@ -73,7 +73,11 @@
 struct packet_s {
 
 	int magic1;		/* for error checking. */
+
 	int seq;		/* unique sequence number for debugging. */
+
+	double release_time;	/* Time stamp in format returned by dtime_now(). */
+				/* When to release from the SATgate mode delay queue. */
 
 #define MAGIC 0x41583235
 
@@ -144,6 +148,8 @@ typedef struct packet_s *packet_t;
 
 
 #ifdef AX25_PAD_C	/* Keep this hidden - implementation could change. */
+
+extern packet_t ax25_new (void);
 
 /*
  * APRS always has one control octet of 0x03 but the more
@@ -273,8 +279,10 @@ typedef struct alevel_s {
 } alevel_t;
 
 
+#ifndef AXTEST
+// TODO: remove this?
 #define AX25MEMDEBUG 1
-
+#endif
 
 
 #if AX25MEMDEBUG	// to investigate a memory leak problem
@@ -310,6 +318,8 @@ extern void ax25_delete (packet_t pp);
 #endif
 
 
+
+
 extern int ax25_parse_addr (int position, char *in_addr, int strict, char *out_addr, int *out_ssid, int *out_heard);
 extern int ax25_check_addresses (packet_t pp);
 
@@ -343,6 +353,9 @@ extern void ax25_set_nextp (packet_t this_p, packet_t next_p);
 extern int ax25_get_dti (packet_t this_p);
 
 extern packet_t ax25_get_nextp (packet_t this_p);
+
+extern void ax25_set_release_time (packet_t this_p, double release_time);
+extern double ax25_get_release_time (packet_t this_p);
 
 extern void ax25_format_addrs (packet_t pp, char *);